# backend/src/swagger.yaml
openapi: 3.0.0
info:
  title: Monster Game API
  version: 1.0.0
  description: API para la gestión de personajes y combates en el juego MonsterGameFight.
servers:
  - url: http://localhost:4000
    description: Servidor local
paths:
...
  /api/messages/send:
    post:
      summary: Enviar un mensaje a otro jugador
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                receiver_id:
                  type: integer
                  example: 3
                subject:
                  type: string
                  example: "Bienvenido al juego"
                body:
                  type: string
                  example: "¡Hola! Espero que disfrutes tu experiencia aquí."
      responses:
        200:
          description: Mensaje enviado correctamente
        400:
          description: Faltan campos obligatorios
        500:
          description: Error interno al enviar el mensaje
  /api/messages:
    get:
      summary: Obtener los mensajes del usuario autenticado
      security:
        - bearerAuth: []
      responses:
        200:
          description: Lista de mensajes obtenida correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        sender_id:
                          type: integer
                        receiver_id:
                          type: integer
                        subject:
                          type: string
                        body:
                          type: string
                        timestamp:
                          type: string
                          format: date-time
                        read:
                          type: boolean
        500:
          description: Error interno al obtener los mensajes
  /api/messages/read/{message_id}:
    post:
      summary: Marcar un mensaje como leído
      security:
        - bearerAuth: []
      parameters:
        - name: message_id
          in: path
          required: true
          schema:
            type: integer
          example: 5
      responses:
        200:
          description: Mensaje marcado como leído correctamente
        400:
          description: El ID del mensaje es obligatorio
        500:
          description: Error interno al marcar el mensaje como leído
  /api/store:
    get:
      summary: Obtener la lista de ítems disponibles en la tienda
      responses:
        200:
          description: Lista de ítems disponibles en la tienda
          content:
            application/json:
              schema:
                type: object
                properties:
                  store:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        name:
                          type: string
                        type:
                          type: string
                          enum: ["weapon", "armor", "accessory"]
                        attack_bonus:
                          type: integer
                        defense_bonus:
                          type: integer
                        price:
                          type: integer
                        rarity:
                          type: string
                          enum: ["common", "rare", "legendary"]
                        level_required:
                          type: integer
        500:
          description: Error interno al obtener la tienda
  /api/store/buy:
    post:
      summary: Comprar un ítem de la tienda
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                item_id:
                  type: integer
                  example: 3
      responses:
        200:
          description: Compra realizada con éxito
        400:
          description: No tienes suficiente oro o el ítem no existe
        404:
          description: Personaje no encontrado
        500:
          description: Error interno al comprar el ítem
  /api/store/sell:
    post:
      summary: Vender un ítem del inventario
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                item_id:
                  type: integer
                  example: 2
      responses:
        200:
          description: Ítem vendido con éxito
        400:
          description: No posees este ítem
        404:
          description: Personaje no encontrado
        500:
          description: Error interno al vender el ítem
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT